+++ tests/test_authorization.py
@@ -98,3 +98,13 @@ class TestAuthorization(unittest.TestCase):
                                      files={"file": "tests/resources/file.png"})
         res_context = auth.apply_authentication(context=context)
         self.assertEqual("11PATHS 2kNhWLEETQ46KWLnAg48 8/fuEv9NLn41ikh96hRHMFGs1ww=", res_context.headers["Authorization"])
+
+    def test_11paths_authentication_json_empty_body_params(self):
+        auth = X11PathsAuthentication(app_id="2kNhWLEETQ46KWLnAg48", secret="lBc4BSeqCGkidJZXictc3yiHbKBS87hjE05YrswJ",
+                                      utc="2017-01-27 08:27:44")
+        context = HttpRequestContext(method="POST", url_path="/ExternalApi/CleanFile",
+                                     renderer=JSONRenderer())
+        res_context = auth.apply_authentication(context=context)
+        self.assertEqual("11PATHS 2kNhWLEETQ46KWLnAg48 atYkLRYJ3b+CXU+GdklyALAr9NE=",
+                         res_context.headers["Authorization"])
+        self.assertNotIn(X_11PATHS_BODY_HASH_HEADER_NAME, res_context.headers)